Feature fix tsolve sync (#35)
* DeltaT sync solve now converge with grad of temp in solution.
Need to check with gradient of specs. Some FluxReg seems to be missing
so it should not work yet ...

* Add FluxReg on Specs for Dn and Dnp1

* DeltaT sync solve now converge with grad of temp in solution.
Need to check with gradient of specs. Some FluxReg seems to be missing
so it should not work yet ...

* Add FluxReg on Specs for Dn and Dnp1

* Fixing mac_sync on species and rhoH.
An update of IAMR diffuse_sync is also required.
Change BL_SPACEDIM in AMREX_SPACEDIM (my own OCD ...) when encountered.

* Fix bug from previous conflict resolution.

* Add a DiffGaussian case where we compute the diffusion of an initial
gaussian perturbation of Temp/Spec. Inputs are provided for single and
two levels cases for both temperature and species. Since no convection
(almost), simulation is performed at a fixed dt.

* Fix 0.5 CN for the diffuse_sync of species.

Co-authored-by: Anne Felden <Enna@lbl.gov>